requirements.txt
# PC-LiquidGAN Requirements
# Install: pip install -r requirements.txt

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Neural ODE solver (REQUIRED for generator)
torchdiffeq>=0.2.3

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-image>=0.19.0

# Scientific computing
numpy>=1.23.0
scipy>=1.9.0
matplotlib>=3.6.0

# Training monitoring
tensorboard>=2.12.0

# Optional: FID score computation
torch-fidelity>=0.3.0

# Optional: experiment tracking
# wandb>=0.15.0
